requirements.txt
# Python 3.9 compatible versions
numpy==1.26.4
scikit-learn==1.5.2
oracledb==2.4.1
cohere==5.11.3
python-dotenv==1.0.1
pypdf